Moving library files to a separate directory
In preparation of divorsing actual bots/scripts from main library
