+++ requirements.txt
@@ -1,0 +1,3 @@
+pywikibot
+MySQL-python
+requests
